Make $Prophet::Replica::PROP_CACHE an object attribute.
I am entirely unconvinced that $PROP_CACHE in the sqlite backend should be a
package variable rather than an object attribute. If you're instantiating
a new handle object, you really probably do want the cache to be empty. If
you want the cache to persist, use the same CLI/handle objects (like the
shell command does). IMHO, all that having a package variable here gets
you is subtle and hard-to-track-down bugs and weird behaviour when
testing.