Annouce project_name via Bonjour
The previous Bonjour annoncement included only the UUID of the source.
That is not very helpfull to decide which source to use.

This change updates the Bonjour annoucement to include both the
project_name and the UUID.

The pull --local code was updated to deal with both new and old servers,
but old clients will not see new servers.

Signed-off-by: Pedro Melo <melo@simplicidade.org>